backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
boto3==1.35.7
opensearch-py==2.6.0
python-dotenv==1.0.1
django-storages==1.14.4
django-filter==24.3
django-redis==5.4.0
uvicorn==0.30.6
gunicorn==22.0.0

